+++ src/main/java/com/nucleus/requestModels/SecondPartial.java
@@ -1,49 +1,49 @@
 package com.nucleus.requestModels;
 
 public class SecondPartial {
-	private int approved;
-	private int notApproved;
-	private int man;
-	private int women;
+	private Long approved;
+	private Long notApproved;
+	private Long man;
+	private Long women;
 	
 	public SecondPartial() {}
 	
-	public SecondPartial(int approved, int notApproved, int man, int women) {
+	public SecondPartial(Long approved, Long notApproved, Long man, Long women) {
 		this.approved = approved;
 		this.notApproved = notApproved;
 		this.man = man;
 		this.women = women;
 	}
 
-	public int getApproved() {
+	public Long getApproved() {
 		return approved;
 	}
 
-	public void setApproved(int approved) {
+	public void setApproved(Long approved) {
 		this.approved = approved;
 	}
 
-	public int getNotApproved() {
+	public Long getNotApproved() {
 		return notApproved;
 	}
 
-	public void setNotApproved(int notApproved) {
+	public void setNotApproved(Long notApproved) {
 		this.notApproved = notApproved;
 	}
 
-	public int getMan() {
+	public Long getMan() {
 		return man;
 	}
 
-	public void setMan(int man) {
+	public void setMan(Long man) {
 		this.man = man;
 	}
 
-	public int getWomen() {
+	public Long getWomen() {
 		return women;
 	}
 
-	public void setWomen(int women) {
+	public void setWomen(Long women) {
 		this.women = women;
 	}
 }
